refacto: moving file to right folder
